# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Open-Voice-Journal API
  version: 0.3.2
  description: API contract for recording ingestion, uploads, provider configuration, and summaries.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/admin/settings/auth:
    get:
      summary: Get auth settings (admin)
      responses:
        '200':
          description: Auth settings loaded
    patch:
      summary: Update auth settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                openSignupEnabled:
                  type: boolean
      responses:
        '200':
          description: Auth settings updated
  /api/v1/admin/openapi-keys:
    get:
      summary: List current admin OpenAPI keys
      responses:
        '200':
          description: Keys listed
    post:
      summary: Create a new admin OpenAPI key (plaintext returned once)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Key created
  /api/v1/admin/openapi-keys/{id}:
    delete:
      summary: Revoke an OpenAPI key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Key revoked
  /api/v1/admin/users/{id}/promote:
    post:
      summary: Promote user to admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User promoted
  /api/v1/admin/users/{id}/demote:
    post:
      summary: Demote admin to user with safeguards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User demoted
        '400':
          description: cannot_self_demote or cannot_demote_last_active_admin
...